app/requirements.txt
flask
mlflow==2.7.1
torch
torchvision
pillow
numpy